Prevent enter key triggering first found button
Previously when an user pressed the enter key on another button than one
of the dual-listbox buttons, the first found button from the
dual-listbox could be triggered. Setting the button type attribute to
'button' prevents this while it is still possible to use the
dual-listbox keys with the enter button while they are focussed.